refactor logic so data returned by makeCalls is already mapped to GraphQL return schema
